+++ @navikt/core/react/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @navikt/ds-react
 
+## 1.5.8
+
+### Patch Changes
+
+- console.log removed
+
+- Updated dependencies []:
+  - @navikt/ds-icons@1.5.8
+
+## 1.5.7
+
+### Patch Changes
+
+- defaultMonth/Year prop date-komponent
+
+- Updated dependencies []:
+  - @navikt/ds-icons@1.5.7
+
 ## 1.5.6
 
 ### Patch Changes